# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Hệ thống Chấm công QR API
  description: |
    API for QR Code Attendance System with GPS verification.

    ## Authentication
    All endpoints require JWT Bearer token except `/auth/login`.

    ## Rate Limiting
    - Authentication: 5 requests/minute
    - Attendance: 10 requests/minute
    - General: 100 requests/minute
  version: 1.0.0
  contact:
    name: API Support
    email: support@company.com
servers:
  - url: http://localhost:8080/api/v1
    description: Development
  - url: https://attendance.company.com/api/v1
    description: Production
...
paths:
...
  /reports/salary:
    get:
      tags: [Reports]
      operationId: getSalaryReport
      summary: Get salary report
      description: Generate salary report (Admin/HR only)
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: integer
        - name: year
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Salary report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalaryReportResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /salary/positions:
    get:
      tags: [Salary Config]
      operationId: getSalaryPositions
      summary: Get all salary positions
      description: Get list of salary positions (Admin/HR only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Position list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SalaryPositionResponse'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Salary Config]
      operationId: createSalaryPosition
      summary: Create salary position
      description: Create new salary position (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalaryPositionRequest'
      responses:
        '201':
          description: Position created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalaryPositionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /salary/positions/{id}:
    put:
      tags: [Salary Config]
      operationId: updateSalaryPosition
      summary: Update salary position
      description: Update salary position (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalaryPositionRequest'
      responses:
        '200':
          description: Position updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalaryPositionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Salary Config]
      operationId: deleteSalaryPosition
      summary: Delete salary position
      description: Delete salary position (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Position deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ========== Salary Experience ==========
  /salary/experience:
    get:
      tags: [Salary Config]
      operationId: getSalaryExperience
      summary: Get all experience tiers
      description: Get list of experience tiers (Admin/HR only)
      security:
        - bearerAuth: []
      parameters:
        - name: positionId
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Experience tier list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SalaryExperienceResponse'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Salary Config]
      operationId: createSalaryExperience
      summary: Create experience tier
      description: Create new experience tier (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalaryExperienceRequest'
      responses:
        '201':
          description: Experience tier created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalaryExperienceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /salary/experience/{id}:
    put:
      tags: [Salary Config]
      operationId: updateSalaryExperience
      summary: Update experience tier
      description: Update experience tier (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalaryExperienceRequest'
      responses:
        '200':
          description: Experience tier updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalaryExperienceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Salary Config]
      operationId: deleteSalaryExperience
      summary: Delete experience tier
      description: Delete experience tier (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Experience tier deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ========== Salary Penalties ==========
  /salary/penalties:
    get:
      tags: [Salary Config]
      operationId: getSalaryPenalties
      summary: Get all penalty rules
      description: Get list of attendance penalty rules (Admin/HR only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Penalty rule list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SalaryPenaltyResponse'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Salary Config]
      operationId: createSalaryPenalty
      summary: Create penalty rule
      description: Create new attendance penalty rule (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalaryPenaltyRequest'
      responses:
        '201':
          description: Penalty rule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalaryPenaltyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /salary/penalties/{id}:
    put:
      tags: [Salary Config]
      operationId: updateSalaryPenalty
      summary: Update penalty rule
      description: Update attendance penalty rule (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalaryPenaltyRequest'
      responses:
        '200':
          description: Penalty rule updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalaryPenaltyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Salary Config]
      operationId: deleteSalaryPenalty
      summary: Delete penalty rule
      description: Delete attendance penalty rule (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Penalty rule deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ========== Salary Bonus ==========
  /salary/bonus:
    get:
      tags: [Salary Config]
      operationId: getSalaryBonus
      summary: Get all bonus configs
      description: Get list of bonus configurations (Admin/HR only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Bonus config list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SalaryBonusResponse'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Salary Config]
      operationId: createSalaryBonus
      summary: Create bonus config
      description: Create new bonus configuration (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalaryBonusRequest'
      responses:
        '201':
          description: Bonus config created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalaryBonusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /salary/bonus/{id}:
    put:
      tags: [Salary Config]
      operationId: updateSalaryBonus
      summary: Update bonus config
      description: Update bonus configuration (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalaryBonusRequest'
      responses:
        '200':
          description: Bonus config updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalaryBonusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Salary Config]
      operationId: deleteSalaryBonus
      summary: Delete bonus config
      description: Delete bonus configuration (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Bonus config deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ========== Employee Salary Assignment ==========
  /salary/assign/position:
    post:
      tags: [Salary Config]
      operationId: assignPosition
      summary: Assign position to employee
      description: Assign salary position to employee (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignPositionRequest'
      responses:
        '201':
          description: Position assigned
        '400':
          $ref: '#/components/responses/BadRequest'
  /salary/assign/experience:
    post:
      tags: [Salary Config]
      operationId: assignExperience
      summary: Assign experience tier to employee
      description: Assign salary experience tier to employee (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignExperienceRequest'
      responses:
        '201':
          description: Experience tier assigned
        '400':
          $ref: '#/components/responses/BadRequest'
  /salary/assign/bonus:
    post:
      tags: [Salary Config]
      operationId: assignBonus
      summary: Assign bonus to employee
      description: Assign salary bonus to employee (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignBonusRequest'
      responses:
        '201':
          description: Bonus assigned
        '400':
          $ref: '#/components/responses/BadRequest'

    delete:
      tags: [Salary Config]
      operationId: unassignBonus
      summary: Remove bonus from employee
      description: Remove salary bonus from employee (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignBonusRequest'
      responses:
        '204':
          description: Bonus removed
  /salary/employee/{userId}:
    get:
      tags: [Salary Config]
      operationId: getEmployeeSalaryDetail
      summary: Get employee salary detail
      description: Get full salary detail for employee (Admin/HR only)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Employee salary detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeSalaryDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ========== Salary Formula ==========
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    SalaryReportResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            month:
              type: integer
            year:
              type: integer
            employees:
              type: array
              items:
                type: object
                properties:
                  userId:
                    type: integer
                  fullName:
                    type: string
                  baseSalary:
                    type: number
                  deductions:
                    type: number
                  totalSalary:
                    type: number
...
    SalaryPositionResponse:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        baseSalary:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    SalaryPositionRequest:
      type: object
      required: [code, name, baseSalary]
      properties:
        code:
          type: string
          minLength: 2
          maxLength: 50
        name:
          type: string
          minLength: 2
          maxLength: 100
        baseSalary:
          type: number
          minimum: 0
        isActive:
          type: boolean
          default: true

    # ========== Salary Experience ==========
    SalaryExperienceResponse:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        positionId:
          type: integer
        positionName:
          type: string
        minYears:
          type: integer
        maxYears:
          type: integer
          nullable: true
        percentIncrease:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    SalaryExperienceRequest:
      type: object
      required: [code, name, positionId, minYears, percentIncrease]
      properties:
        code:
          type: string
          minLength: 2
          maxLength: 50
        name:
          type: string
          minLength: 2
          maxLength: 100
        positionId:
          type: integer
        minYears:
          type: integer
          minimum: 0
        maxYears:
          type: integer
          nullable: true
        percentIncrease:
          type: number
          minimum: 0
          maximum: 200
        isActive:
          type: boolean
          default: true

    # ========== Salary Penalty ==========
    SalaryPenaltyResponse:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        penaltyType:
          type: string
          enum: [LATE, ABSENT]
        minMinutes:
          type: integer
        maxMinutes:
          type: integer
          nullable: true
        penaltyAmount:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    SalaryPenaltyRequest:
      type: object
      required: [code, name, penaltyType, minMinutes, penaltyAmount]
      properties:
        code:
          type: string
          minLength: 2
          maxLength: 50
        name:
          type: string
          minLength: 2
          maxLength: 100
        penaltyType:
          type: string
          enum: [LATE, ABSENT]
        minMinutes:
          type: integer
          minimum: 0
        maxMinutes:
          type: integer
          nullable: true
        penaltyAmount:
          type: number
          minimum: 0
        isActive:
          type: boolean
          default: true

    # ========== Salary Bonus ==========
    SalaryBonusResponse:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        amount:
          type: number
        description:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    SalaryBonusRequest:
      type: object
      required: [code, name, amount]
      properties:
        code:
          type: string
          minLength: 2
          maxLength: 50
        name:
          type: string
          minLength: 2
          maxLength: 100
        amount:
          type: number
          minimum: 0
        description:
          type: string
        isActive:
          type: boolean
          default: true

    # ========== Employee Assignment ==========
    AssignPositionRequest:
      type: object
      required: [userId, positionId]
      properties:
        userId:
          type: integer
        positionId:
          type: integer
    AssignExperienceRequest:
      type: object
      required: [userId, experienceId]
      properties:
        userId:
          type: integer
        experienceId:
          type: integer
    AssignBonusRequest:
      type: object
      required: [userId, bonusId]
      properties:
        userId:
          type: integer
        bonusId:
          type: integer

    # ========== Employee Salary Detail ==========
    EmployeeSalaryDetailResponse:
      type: object
      properties:
        userId:
          type: integer
        fullName:
          type: string
        position:
          type: object
          properties:
            id:
              type: integer
            code:
              type: string
            name:
              type: string
            baseSalary:
              type: number
        experience:
          type: object
          nullable: true
          properties:
            id:
              type: integer
            code:
              type: string
            name:
              type: string
            percentIncrease:
              type: number
        bonuses:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              code:
                type: string
              name:
                type: string
              amount:
                type: number
        yearsOfWork:
          type: integer
        totalSalary:
          type: number
          description: Calculated by formula

    # ========== Salary Formula ==========
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...